tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/to_keyframe_omitted.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - 100% or "to" keyframe not specified</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#keyframes">
   <meta name="flags" content="dom">
   <meta name="assert" content='If a 100% or "to" keyframe is not specified,
   then the user agent constructs a 100% keyframe using the computed values of
   the properties being animated'>
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         0% {
            left: 100px;
         }
         75% {
            left: 80px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
         left: 0;
      }
      .test {
         animation: slide 1s linear forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var passed;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            var passed = (getComputedStyle(div, null).left == '0px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         });
         setTimeout(function() {
            /* ISSUE: David L. Baron thinks such changes should be picked up
             * http://lists.w3.org/Archives/Public/www-style/2011Sep/0423.html
             */
            //div.style.left = '200px';
         }, 500);
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should slide straight leftwards for one second, slowly
   at first and then faster.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
